requirements.txt
annotated-types==0.6.0
anyio==4.3.0
click==8.1.7
diff-match-patch==20181111
fastapi==0.110.0
h11==0.14.0
httptools==0.6.1
idna==3.6
loguru==0.7.2
pydantic==2.6.3
pydantic-settings==2.2.1
pydantic_core==2.16.3
python-dotenv==1.0.1
PyYAML==6.0.1
quill-delta==1.0.3
sniffio==1.3.1
starlette==0.36.3
typing_extensions==4.10.0
uvicorn==0.28.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
python-multipart==0.0.9
pillow==10.3.0
